feat: single loading schema module
