+++ AGENTS.md
@@ -32,12 +32,12 @@ Spine is a configuration-first ingestion framework. Treat it like production pip
 - Do not commit operator-local config, secrets, or internal-only endpoints. Public templates live in `config/defaults.example.yml` and `config/examples/`.
 - If you change config schema, also update the relevant examples and docs in `config/README.md` and `docs/configuration/`.
 - Respect `CONFIG_PATH` behavior in `src/config/settings.py` when changing config resolution.
-- Loading identity strings (object-store bucket/container/account labels, slash-trimmed `prefix`, REST `base_url` trailing slashes) are normalized during Pydantic validation. Shared helpers live in `src/config/loading_fields.py`. Spark base URIs use `loading_base_uri(LoadingConfig)` in `src/loader/object_store.py`: it expects **validated** config only and does not re-normalize identity fields (resolve `storage_root` to `file:` URIs still happens there). CLI-only inputs—for example operator `s3://` URIs for config promotion—normalize once at parse time in `src/utils/s3_config_push.parse_s3_uri`, not via pipeline loading validators. Path joining for Spark segments and HTTP URL composition remain in loader/service helpers.
+- Loading identity strings (object-store bucket/container/account labels, slash-trimmed `prefix`, REST `base_url` trailing slashes) are normalized during Pydantic validation. Destination aliases and those normalizers live in `src/config/loading_schema.py`. Spark base URIs use `loading_base_uri(LoadingConfig)` in `src/loader/object_store.py`: it expects **validated** config only and does not re-normalize identity fields (resolve `storage_root` to `file:` URIs still happens there). CLI-only inputs—for example operator `s3://` URIs for config promotion—normalize once at parse time in `src/utils/s3_config_push.parse_s3_uri`, not via pipeline loading validators. Path joining for Spark segments and HTTP URL composition remain in loader/service helpers.
 
 ## Extending Spine
 
 - New auth behavior: extend the service/auth configuration model and validation together.
-- New destination: add a loader under `src/loader/`, register it in `src/loader/loader_factory.py` under the **canonical** destination id (see `src/config/loading_destinations.py` for object-store destinations, aliases, and `normalize_loading_destination`), and document any new config. If Spark needs filesystem or connector wiring, extend `src/config/config_spark.py` (`SparkSessionConf`), `src/config/spark_runtime.py`, and `defaults.spark_runtime` in the same change set so behavior stays configuration-first.
+- New destination: add a loader under `src/loader/`, register it in `src/loader/loader_factory.py` under the **canonical** destination id (see `src/config/loading_schema.py` for object-store destinations, aliases, and `normalize_loading_destination`), and document any new config. If Spark needs filesystem or connector wiring, extend `src/config/config_spark.py` (`SparkSessionConf`), `src/config/spark_runtime.py`, and `defaults.spark_runtime` in the same change set so behavior stays configuration-first.
 - New source/service type: implement it under `src/service/`, wire it through the factory/config models, and avoid bypassing planner/handler flow. For a new **relational database** kind that shares the same table/query extract and request-context rules, add its `SourceType` to `is_database_source_type()` in `src/config/config_models.py` (that function is the only database-kind predicate; call it from planner, handler, and validators rather than re-listing types).
 - New transformation or collection behavior should fit the existing parser/collector split rather than being embedded ad hoc in services.
